refactor: 모델 객체 domain 패키지로 이동
