+++ src/main/java/roomescape/controller/ReservationController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.Reservation;
 import roomescape.controller.request.ReservationRequest;
+import roomescape.model.Reservation;
 
 @RestController
 public class ReservationController {
